+++ spec/factories/exercise_tags.rb
@@ -1,0 +1,6 @@
+FactoryBot.define do
+  factory :exercise_tag do
+    association :exercise
+    association :tag
+  end
+end