+++ spec/factories/muscle_group_translations.rb
@@ -1,0 +1,7 @@
+FactoryBot.define do
+  factory :muscle_group_translation do
+    association :muscle_group
+    locale { "en" }
+    sequence(:name) { |n| "Muscle Group #{n}" }
+  end
+end