+++ spec/factories/tag_translations.rb
@@ -1,0 +1,7 @@
+FactoryBot.define do
+  factory :tag_translation do
+    association :tag
+    locale { "en" }
+    sequence(:name) { |n| "Tag #{n}" }
+  end
+end